// remix/src/types.ts
export interface ProjectAsset {
  id: string;
  name: string;
  fileName: string;
  type: "drawable" | "mipmap" | "font" | "raw" | "asset";
  format: "png" | "webp" | "svg" | "ttf" | "json" | "mp3" | "mp4";
  size: string;
  url?: string;
  targetResDir: string;
}